rename childCombined file and update year logic
