+++ services/ui-src/src/shared/Qualifiers/validations/adultMedicaid.ts
@@ -1,20 +1,13 @@
+import { featuresByYear } from "utils/featuresByYear";
 import {
   ACSMQualifierForm,
   DeliverySystem,
 } from "../../types/TypeQualifierForm";
 
-const validate21To64EqualsToOneHundredPercent = (
-  data: ACSMQualifierForm,
-  year: number | undefined
-) => {
-  // Extract year from URL
-  if (!year && typeof window !== "undefined") {
-    const pathYear = window.location.pathname.match(/\/(\d{4})\//)?.[1];
-    year = pathYear ? parseInt(pathYear) : undefined;
-  }
-
+const validate21To64EqualsToOneHundredPercent = (data: ACSMQualifierForm) => {
   const values = data["PercentageEnrolledInEachDeliverySystem"];
   const errorArray: any[] = [];
+
   const total21To64Percent = values?.reduce(
     (acc: number, curr: DeliverySystem) => {
       return acc + parseFloat(curr.TwentyOneToSixtyFour || "0");
@@ -37,7 +30,14 @@ const validate21To64EqualsToOneHundredPercent = (
     (total64PlusPercent < 99 || total64PlusPercent > 101) &&
     total64PlusPercent !== 0;
 
-  if (year === 2025) {
+  if (has21To64ZeroError) {
+    errorArray.push({
+      errorLocation: "Delivery System",
+      errorMessage: "Entries for Ages 21 to 64 column must have values",
+    });
+  }
+
+  if (featuresByYear.lessSpecificQualifierValidationLanguage) {
     // For 2025, show one generic message if any validation fails
     if (has21To64ZeroError || has21To64TotalError || has64PlusTotalError) {
       errorArray.push({
@@ -47,12 +47,6 @@ const validate21To64EqualsToOneHundredPercent = (
     }
   } else {
     // For other years, show specific messages for each validation
-    if (has21To64ZeroError) {
-      errorArray.push({
-        errorLocation: "Delivery System",
-        errorMessage: "Entries for Ages 21 to 64 column must have values",
-      });
-    }
     if (has21To64TotalError) {
       errorArray.push({
         errorLocation: "Delivery System",
