+++ services/ui-src/src/shared/Qualifiers/validations/childMedicaid.ts
@@ -1,20 +1,13 @@
+import { featuresByYear } from "utils/featuresByYear";
 import {
   CCSMQualifierForm,
   DeliverySystem,
 } from "../../types/TypeQualifierForm";
 
-const validate21To64EqualsToOneHundredPercent = (
-  data: CCSMQualifierForm,
-  year: number | undefined
-) => {
-  // Extract year from URL
-  if (!year && typeof window !== "undefined") {
-    const pathYear = window.location.pathname.match(/\/(\d{4})\//)?.[1];
-    year = pathYear ? parseInt(pathYear) : undefined;
-  }
-
+const validate21To64EqualsToOneHundredPercent = (data: CCSMQualifierForm) => {
   const values = data["PercentageEnrolledInEachDeliverySystem"];
   const errorArray: any[] = [];
+
   const totalUnder21MedicaidPercent = values?.reduce(
     (acc: number, curr: DeliverySystem) => {
       return acc + parseFloat(curr.UnderTwentyOne || "0");
@@ -29,18 +22,23 @@ const validate21To64EqualsToOneHundredPercent = (
     });
   }
 
-  if (
-    totalUnder21MedicaidPercent > 0 &&
-    (totalUnder21MedicaidPercent < 99 || totalUnder21MedicaidPercent > 101)
-  ) {
-    errorArray.push({
-      errorLocation: "Delivery System",
-      errorMessage:
-        year === 2025
-          ? "Entries for column must total 100"
-          : "Entries for Under Age 21 Medicaid column must total 100",
-    });
-  }
+  const hasUnder21MedicaidTotalError =
+    (totalUnder21MedicaidPercent > 0 && totalUnder21MedicaidPercent < 99) ||
+    totalUnder21MedicaidPercent > 101;
+
+  if (hasUnder21MedicaidTotalError)
+    if (featuresByYear.lessSpecificQualifierValidationLanguage) {
+      errorArray.push({
+        errorLocation: "Delivery System",
+        errorMessage: "Entries for column must total 100",
+      });
+    } else {
+      errorArray.push({
+        errorLocation: "Delivery System",
+        errorMessage: "Entries for Under Age 21 Medicaid column must total 100",
+      });
+    }
+
   return errorArray.length ? errorArray : [];
 };
