+++ services/ui-src/src/shared/Qualifiers/validations/index.ts
@@ -1,7 +1,7 @@
 export * from "./adult";
 export * from "./adultChip";
 export * from "./adultMedicaid";
-export * from "./childCombined";
+export * from "./child";
 export * from "./childCHIP";
 export * from "./childMedicaid";
 export * from "./healthHome";
